Moved AccountDetailsContainer to containers folder
