+++ app/components/AccountDetails.jsx
@@ -4,7 +4,9 @@ export default class AccountDetails extends React.Component {
   //Needs currentUser for the user info like name and email
   //Needs to grab all addresses of currentUser as currentUserAddresses
   //Needs a button for changing password, viewing order history and viewing reviews
+
   render() {
+    //console.log(this.props.selectUser);
     return (
       <div>
         <label>Name:</label>
