+++ app/main.jsx
@@ -7,10 +7,12 @@ import {connect, Provider} from 'react-redux'
 import store from './store'
 
 import AppContainer from './components/AppContainer'
-import AccountDetailsContainer from './components/AccountDetailsContainer'
+import AccountDetailsContainer from './containers/AccountDetailsContainer'
 import Jokes from './components/Jokes'
 import Login from './components/Login'
 import WhoAmI from './components/WhoAmI'
+
+import { fetchUser } from './reducers/user';
 
 
 const ExampleApp = connect(
@@ -30,7 +32,7 @@ render(
     <Router history={browserHistory}>
       <Route path="/" component={AppContainer}>
       </Route>
-      <Route path="/account" component={AccountDetailsContainer}/>
+      <Route path="/account" component={AccountDetailsContainer} onEnter={fetchUser}/>
     </Router>
   </Provider>,
   document.getElementById('main')
